backend/requirements.txt
fastapi == 0.115.12
uvicorn == 0.34.2
langchain == 0.3.23
langchain-community == 0.3.21
langchain-openai == 0.3.14
langgraph == 0.3.31
openai == 1.75.0
python-dotenv == 1.1.0
boto3==1.38.6